separated app server and socket
app is now loaded in socket, both socket and http servers are created in socket and listened for in server file...
