+++ src/app.ts
@@ -7,8 +7,8 @@ const app = express();
 app.use(cors({
     origin: 'http://localhost:3000',
     credentials: true,
-    
 }))
+
 app.use(cookieParser())
 app.use(express.json())
 app.use(express.urlencoded({extended: false}))
