+++ src/socket/index.ts
@@ -1,7 +1,19 @@
-import { io } from "../server";
+import { Server as socketio } from "socket.io";
 import { decodeToken, jwtMiddleware } from "../middlewares/jwt";
 import { writeMessageToDB } from "../handlers/messages";
 import Logger from "../logger/logger";
+import http from "http";
+import app from "../app";
+
+export const server = http.createServer(app);
+
+export const io = new socketio(server, {
+  cors: {
+    origin: "http://localhost:3000",
+    allowedHeaders: ["cookies"],
+    credentials: true,
+  },
+});
 
 // Socket.io
 io.use(async (socket, next) => {
@@ -21,20 +33,26 @@ io.use(async (socket, next) => {
 
 io.on("connection", (socket) => {
   Logger.info("Socket connected");
-  socket.emit("message", "Welcome to the chat");
 
+  // ON JOIN ROOM JOIN SAID SOCKET TO SAID ROOM
   socket.on("joinRoom", async (room_id) => {
     await socket.join(room_id);
     Logger.info(`Socket ${socket.id} joined room ${room_id}`);
-
-    socket.to(room_id).emit("update", "successfully joined!");
   });
 
+  // ON MESSAGE: UPDATE EVERYONE ELSE
   socket.on("message", async (data) => {
     Logger.info(`Socket ${socket.id} sent message ${data}`);
 
-    await socket.to(data.room).emit("update", data.message);
-    const handlerReturn = await writeMessageToDB(data);
-    console.log(handlerReturn);
+    const newData = {
+      eventName: "update",
+      content: data.content,
+      sentBy: data.sentBy,
+      createdAt: new Date(),
+      room_id: data.room_id,
+    };
+
+    await socket.to(data.room_id).emit("update", newData);
+    await writeMessageToDB(newData);
   });
 });
